Clean up tag code, replace error event
